TURMERIC-1600: create turmericsoa-security common folder
The following handlers have been added to turmerisoa-security folder:
* AuthenticationServiceConsumer
* AuthorizationServiceConsumer
* PolicyEnforncementConsumer
* GroupMembershipServiceConsumer
* PolicyServiceConsumer

These and their related interfaces are in the common folder for use
by various web apps.  Otherwise they have to be bundled with the
WAR file for each.   Downside to this is that they will share a
common configuration so make sure that you want all the services
talking to the same remote services or local instances.